Initial support for Sonarqube 8.1
Sonarqube 8.1 removed the concepts of short-lived and long-lived branches and replaced it with a singular `branch` concept. To allow the plugin to continue to support older versions of Sonarqube, the original enum values are checked for and used where they exist, with the new enum value being used where the older value cannot be found.